+++ src/app/community-builds/CommunityBuilds.tsx
@@ -146,7 +146,7 @@ export function CommunityBuildList({ itemsPerPage = 8 }: Props) {
       >
         <ul
           role="list"
-          className="my-4 grid grid-cols-1 gap-6 sm:grid-cols-2 lg:grid-cols-4"
+          className="my-4 grid grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-2 lg:grid-cols-4"
         >
           {builds.map((build) => (
             <BuildCard